Renamed "functional" tests to "integration" tests
