+++ test/Github/Tests/Integration/RateLimitTest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Github\Tests\Functional;
+namespace Github\Tests\Integration;
 
 /**
- * @group functional
+ * @group integration
  */
 class RateLimitTest extends TestCase
 {
